requirements.txt
numpy
Flask==0.11.1
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp36-cp36m-linux_x86_64.whl
torchvision
pillow
gunicorn
